predict

+++ app.py
@@ -228,7 +228,7 @@ def process_user_input(user_df, encoders, scalers, feature_pairs):
         if max_val != min_val:
             user_df[new_col_name] = 2 * ((user_df[new_col_name] - min_val) / (max_val - min_val)) - 1
         else:
-            user_df[new_col_name] =  int(str(1 + (0 * len(str(user_df[new_col_name])))))
+            user_df[new_col_name] =  float(str(1 + (0 * len(str(user_df[new_col_name])))))
 
     return user_df
